+++ src/Modules/Region/components/Pin.vue
@@ -20,7 +20,7 @@
 
           <LeafletLocationPicker
             :zoom="6"
-            :coordinates="[lat, lon]"
+            :coordinates="{lat: inlat, lon: inlon}"
             :icon="locationPickerIcon"
             @coordinates-change="updateCoordinates"
           />
@@ -113,8 +113,8 @@ export default {
       this.isLoading = false
     },
     updateCoordinates (coords) {
-      this.inlat = coords[0]
-      this.inlon = coords[1]
+      this.inlat = coords.lat
+      this.inlon = coords.lon
     },
   },
 }