+++ tests/acceptance/SettingsCest.php
@@ -136,19 +136,19 @@ class SettingsCest
         $I->login($this->foodsaver['email']);
         $I->amOnPage('/?page=settings&sub=general');
         $I->waitForPageBody();
-        $I->fillField('#addresspicker', $address);
-        $I->waitForElementVisible('#addresspicker_listbox');
-        $I->click("//*[@id='addresspicker_listbox']//*[contains(text(), 'Teststraße 1')]");
+        $I->fillField('#searchinput', $address);
+        $I->waitForElementVisible('#searchinput_listbox');
+        $I->click("//*[@id='searchinput_listbox']//*[contains(text(), 'Teststraße 1')]");
         $I->click('Speichern');
         $I->waitForPageBody();
 
         $I->amOnPage('/?page=settings&sub=general');
         $I->waitForPageBody();
-        $I->seeInField('#anschrift', 'Teststraße 1');
-        $I->seeInField('#plz', '37073');
-        $I->seeInField('#ort', 'Teststadt');
-        $I->assertEqualsWithDelta($I->grabValueFrom('#lat'), 51.0, 0.001);
-        $I->assertEqualsWithDelta($I->grabValueFrom('#lon'), 9.0, 0.001);
+        $I->seeInField('#input-street', 'Teststraße 1');
+        $I->seeInField('#input-postal', '37073');
+        $I->seeInField('#input-city', 'Teststadt');
+        $I->assertEqualsWithDelta($I->grabValueFrom('input[name=lat]'), 51.0, 0.001);
+        $I->assertEqualsWithDelta($I->grabValueFrom('input[name=lon]'), 9.0, 0.001);
     }
 
     private function createSelector(string $field): string